requirements.txt
folium
numpy
pillow